Extend FLEXPART result query to contain offline run
Previously only the path to the stashed results from the main run and the post-processing step was returned.

Query had to be extended to cointain potentially offline nested run. The stash folder of the later will now be retrieved as well.

Output of csv tables were extended for offline nested stash address as well.

Some renaming to clarify variable/column names.